requirements.txt
fastapi
uvicorn
openai-whisper
python-dotenv
requests
moviepy<2.0
pydub
google-genai
redis
python-multipart
